api/requirements.txt
streamlit>=1.28.0
psycopg2>=2.9.10
# Core ML dependencies
torch>=2.6.0
transformers>=4.40.0
tokenizers>=0.15.0

# Google Gemini API
google-generativeai>=0.8.0

# FAISS and embeddings
faiss-cpu>=1.7.4
FlagEmbedding>=1.2.0
sentence-transformers>=2.7.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

# Development and testing
pytest>=8.2.0
pytest-cov>=5.0.0
black>=24.4.0
isort>=5.13.0
